Fix server-side logging
The PrintWriter is now explicitly flushed whenever a line is written to
the log. It can now also log messages when they were sent encrypted
(no, it was not the encryption stopping us in the first place, it only
checked for ChatMessage and not ChatEncryptedMessage to log) and
logging is now not an inherent part of broadcasting.