Add start of multi-unit system
This overhauls things pretty fundamentally, and it's WIP, so it's in a
new branch.

Now, a weather station will have its "raw" values from the API wrapped
inside WeatherAttribute objects, instead of direct attributes of the
WeatherStation object. WeatherAttribute objects have a list of
WeatherAttributeShownValue objects, which are intended for "nice"
display (including, for example, changing "-" to "n/a"). This way, an
attribute can be displayed in multiple units, the temperature can be
shown in any combination of the units already.